% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_extrude_app.R
\name{example_extrude_app}
\alias{example_extrude_app}
\title{Render Example Extrude App}
\usage{
example_extrude_app()
}
\value{
An interactive shiny app of BC's natural resource regions
}
\description{
Render Example Extrude App
}
\examples{
example_extrude_app()
}
