% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_holo_app.R
\name{example_holo_app}
\alias{example_holo_app}
\title{Render Example Holo App}
\usage{
example_holo_app()
}
\value{
An interactive shiny app of Vancouver's download LiDAR
}
\description{
Render Example Holo App
}
\examples{
example_holo_app()
}
