% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrudePlotOutput.R
\name{extrudePlotOutput}
\alias{extrudePlotOutput}
\title{Display a rendered Three.js Extrude Plot}
\usage{
extrudePlotOutput(
  outputId = NULL,
  height = "500px",
  width = "100\%",
  style = "border:none;"
)
}
\arguments{
\item{outputId}{Unique ID to link render and output functions}

\item{height}{Height of widget; can use 'px' or '%'}

\item{width}{Width of widget; can use 'px' or '%'}

\item{style}{Additional CSS style arguments}
}
\value{
Widget for UI of Shiny app, tied to a renderExtrudePlot by shared outputId.
}
\description{
Display a rendered Three.js Extrude Plot
}
