% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holoMapOutput.R
\name{holoMapOutput}
\alias{holoMapOutput}
\title{Display a rendered Three.js Holo Map}
\usage{
holoMapOutput(
  outputId = NULL,
  height = "500px",
  width = "100\%",
  style = "border:none;"
)
}
\arguments{
\item{outputId}{Unique ID to link render and output functions}

\item{height}{Height of widget; can use 'px' or '%'}

\item{width}{Width of widget; can use 'px' or '%'}

\item{style}{Additional CSS style arguments}
}
\value{
Widget for UI of Shiny app, tied to a renderHoloMap by shared outputId.
}
\description{
Display a rendered Three.js Holo Map
}
