% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderExtrudePlot.R
\name{renderExtrudePlot}
\alias{renderExtrudePlot}
\title{Render a Three.js Extrude Plot widget}
\usage{
renderExtrudePlot(
  outputId = NULL,
  dat,
  label_col = NULL,
  height_col = NULL,
  session
)
}
\arguments{
\item{outputId}{Unique ID to link render and output functions}

\item{dat}{Data to use in extrude plot}

\item{label_col}{Column name to use to label extrudes}

\item{height_col}{Column name that should be used to inform height of extrudes}

\item{session}{Shiny session; should just say session = session}
}
\value{
One half of a Shiny-compatible render and output combo. You don't need to assign this to Shiny's output list
}
\description{
Render a Three.js Extrude Plot widget
}
