% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderHoloMap.R
\name{renderHoloMap}
\alias{renderHoloMap}
\title{Render a Three.js Holo Map widget}
\usage{
renderHoloMap(outputId = NULL, dat, session)
}
\arguments{
\item{outputId}{Unique ID to link render and output functions}

\item{dat}{Data to use in extrude plot}

\item{session}{Shiny session; should just say session = session}

\item{height_col}{Column name that should be used to inform height of holographic pillars}
}
\value{
One half of a Shiny-compatible render and output combo. You don't need to assign this to Shiny's output list
}
\description{
Render a Three.js Holo Map widget
}
